refactor: restructure ApplicationCard component and extract related functionalities into separate components and hooks
